Add request spec for favorite

+++ db/migrate/20210223203437_create_favorites.rb
@@ -2,7 +2,7 @@ class CreateFavorites < ActiveRecord::Migration[6.1]
   def change
     create_table :favorites do |t|
       t.string :name
-      t.integer :priority
+      t.bigint :priority
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
